fix: create firebase-config and move it
