+++ src/firebase-config.ts
@@ -1,0 +1,11 @@
+const firebaseConfig = {
+  apiKey: "AIzaSyBvEKd-lv2_xZKlVpVNXwesxIDIJU--UCw",
+  authDomain: "daegyeome-site.firebaseapp.com",
+  projectId: "daegyeome-site",
+  storageBucket: "daegyeome-site.appspot.com",
+  messagingSenderId: "524135997669",
+  appId: "1:524135997669:web:5f83bbb1b2ded804c5b40c",
+  measurementId: "G-DM8X5N4C0K"
+};
+
+export default firebaseConfig;
